Created a component to display the work release form using a popover
